requirements.txt
pandas
pyyaml


#pip install -r .\requirements.txt
#in root project